change startValue parameter of InverseCdf::set_defaults from int to double
This is to avoid data loss during an unnecessary type conversion.

+++ src/InverseCdf.cpp
@@ -18,7 +18,7 @@ void InverseCdf::apply_exponent(double value)
 }
 
 // set defaults if not present in parameter file
-void InverseCdf::set_defaults(int startValue)
+void InverseCdf::set_defaults(double startValue)
 {
 	cdf_values_[CDF_RES] = startValue;
 	for (int i = 0; i < CDF_RES; i++)
